Add checking user auth

+++ src/main/java/com/shinobicoders/teamcodeapi/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.shinobicoders.teamcodeapi.controller;
 
 import com.shinobicoders.teamcodeapi.model.User;
+import com.shinobicoders.teamcodeapi.service.AuthService;
 import com.shinobicoders.teamcodeapi.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +17,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserController {
     private final UserService userService;
+    private final AuthService authService;
 
     @GetMapping
     public ResponseEntity<List<User>> getAllUsers() {
@@ -41,6 +43,11 @@ public class UserController {
 
     @PutMapping("/{id}")
     public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userDetails) {
+
+        if(! authService.authorizeUser(id)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         User updatedUser = userService.updateUser(id, userDetails);
 
         if (updatedUser == null) {
@@ -54,6 +61,10 @@ public class UserController {
     public ResponseEntity<?> deleteUser(@PathVariable Long id) {
         // todo: check auth
         // todo: test what happens if user does not exist
+
+        if(! authService.authorizeUser(id)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
 
         userService.deleteUser(id);
         return new ResponseEntity<>(HttpStatus.OK);
